% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_njogis.R
\name{batch_geocode_addresses}
\alias{batch_geocode_addresses}
\title{Batch geocode addresses}
\usage{
batch_geocode_addresses(df, id, street, city, state, zip, crs = 4326)
}
\arguments{
\item{df}{dataframe with addresses to be geocoded}

\item{id}{primary key that uniquely identifies rows}

\item{street}{street address column}

\item{city}{city column}

\item{state}{state column}

\item{zip}{zip code column, expects a 5-digit zip code}

\item{crs}{coordinate reference system to use for output}
}
\value{
an sf object containing geocoding results
}
\description{
Batch geocode addresses
}
