% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapefiles.R
\name{get_county_bounds}
\alias{get_county_bounds}
\title{Download county boundaries}
\usage{
get_county_bounds(crs = 4326)
}
\arguments{
\item{crs}{coordinate reference system code. This is passed to \code{sf::st_transform()}}
}
\value{
an sf object with boundaries for New Jersey counties
}
\description{
Download county boundaries
}
