% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapefiles.R
\name{get_muni_bounds}
\alias{get_muni_bounds}
\title{Download municipal boundaries}
\usage{
get_muni_bounds(crs = 4326)
}
\arguments{
\item{crs}{coordinate reference system code. This is passed to \code{sf::st_transform()}}
}
\value{
an sf object with boundaries for all New Jersey municipalities
}
\description{
Download municipal boundaries
}
