% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapefiles.R
\name{get_state_bounds}
\alias{get_state_bounds}
\title{Download state boundaries}
\usage{
get_state_bounds(crs = 4326)
}
\arguments{
\item{crs}{coordinate reference system code. This is passed to \code{sf::st_transform()}}
}
\value{
an sf object with boundaries for the State of New Jersey
}
\description{
Download state boundaries
}
