% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_njogis.R
\name{reverse_geocode}
\alias{reverse_geocode}
\title{Reverse geocode a set of coordinates}
\usage{
reverse_geocode(lng, lat, distance = NULL, crs = 4326)
}
\arguments{
\item{lng}{Longitude}

\item{lat}{Latitude}

\item{distance}{Max distance to search around coordinates (in feet)}

\item{crs}{Coordinate reference system code of points}
}
\value{
Place name and address corresponding to coordinates
}
\description{
Reverse geocode a set of coordinates
}
\examples{
reverse_geocode(-74.44513, 40.49297)
}
